requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Django==4.2.4
djangorestframework==3.14.0
exceptiongroup==1.1.2
fastapi==0.100.1
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
